refactor: migrate ArrowArrays to record/ and remove useless errors
